requirements.txt
numpy
pandas
matplotlib
streamlit==0.89.0
Pillow==8.4.0
scikit-learn==0.24.2 